api cleanup

+++ Backend/ReQuests.Api/ReQuests.Api/Controllers/UsersController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ReQuests.Api.Extensions;
 using ReQuests.Domain.Dtos.User;
 
 namespace ReQuests.Api.Controllers;
@@ -12,20 +13,21 @@ public class UsersController : ExtendedControllerBase
 {
 	private readonly IUsersService _usersService;
 
-	public UsersController(IUsersService usersService)
+	public UsersController( IUsersService usersService )
 	{
 		_usersService = usersService;
 	}
 
 	[HttpGet]
-	[Authorize( Roles = "Admin" )]
+	[Authorize( Roles = Constants.Auth.SuperAdminRole )]
 	public async Task<ActionResult<GetUserDto[]>> GetUsers()
 	{
 		return await _usersService.GetUsersAsync();
 	}
 
 	[HttpGet( "{uuid}" )]
-	public async Task<ActionResult<GetUserDto>> GetUser(string uuid)
+	[Authorize( Roles = Constants.Auth.SuperAdminRole )]
+	public async Task<ActionResult<GetUserDto>> GetUser( string uuid )
 	{
 		var user = await _usersService.GetUserAsync( uuid );
 		if ( user is null )
@@ -38,7 +40,7 @@ public class UsersController : ExtendedControllerBase
 
 	[HttpPost]
 	[AllowAnonymous]
-	public async Task<ActionResult<GetUserDto>> CreateUser([FromBody] CreateUserDto dto)
+	public async Task<ActionResult<GetUserDto>> CreateUser( [FromBody] CreateUserDto dto )
 	{
 		GetUserDto user;
 		try
@@ -57,8 +59,49 @@ public class UsersController : ExtendedControllerBase
 	}
 
 	[HttpDelete( "{uuid}" )]
-	public async Task<IActionResult> DeleteUser(string uuid)
+	[Authorize( Roles = Constants.Auth.SuperAdminRole )]
+	public async Task<IActionResult> DeleteUser( string uuid )
 	{
+		try
+		{
+			await _usersService.DeleteUserAsync( uuid );
+		}
+		catch ( NotFoundException )
+		{
+			return NotFound();
+		}
+
+		return NoContent();
+	}
+
+
+	[HttpGet( "me" )]
+	public async Task<ActionResult<GetUserDto>> GetCurrentUser()
+	{
+		var uuid = User.GetUuid();
+		if ( uuid is null )
+		{
+			return InternalServerError( "Error occured" );
+		}
+
+		var user = await _usersService.GetUserAsync( uuid );
+		if ( user is null )
+		{
+			return NotFound();
+		}
+
+		return user;
+	}
+
+	[HttpDelete( "me" )]
+	public async Task<IActionResult> DeleteCurrentUser()
+	{
+		var uuid = User.GetUuid();
+		if ( uuid is null )
+		{
+			return InternalServerError( "Error occured" );
+		}
+
 		try
 		{
 			await _usersService.DeleteUserAsync( uuid );
